# internal/httpserver/handlers/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hostyt Proxy Gateway API",
    "version": "0.1.0",
...
  },
  "servers": [
    {
      "url": "https://proxy.host.yt",
      "description": "Production"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key from `/admin/api-keys` or `/app/api-keys`."
      },
      "nodeAuthToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Per-node agent token. Also accepted as `?node_token=<token>`."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "not found" }
        }
      },
...
      "WgPeerStats": {
        "type": "object",
        "description": "Per-peer stats parsed from `wg show <iface> dump`.",
        "required": ["pubkey"],
        "properties": {
          "pubkey": { "type": "string", "example": "abc123...==" },
          "last_handshake": { "type": "integer", "format": "int64", "description": "Unix epoch of last WireGuard handshake", "example": 1719216000 },
          "rx_bytes": { "type": "integer", "format": "int64", "example": 1048576 },
          "tx_bytes": { "type": "integer", "format": "int64", "example": 524288 },
          "endpoint": { "type": "string", "description": "Observed remote endpoint", "example": "203.0.113.5:51820" }
        }
      },
      "NodeDiagnostics": {
        "type": "object",
        "description": "Node-level forwarding diagnostics. All fields are optional — older agents omit them.",
        "properties": {
          "ip_forward_enabled": { "type": "boolean" },
          "forward_policy_drop_detected": { "type": "boolean" },
          "docker_rules_installed": { "type": "boolean" },
          "firewall_backend": { "type": "string", "example": "nftables" },
          "mtu": { "type": "integer", "example": 1420 },
          "listen_port": { "type": "string", "example": "51821" },
          "last_setup_error": { "type": "string", "example": "" },
          "wstunnel_healthy": {
            "type": "boolean",
            "description": "True when wstunnel sidecar is running and bridge→host iptables rule accepted. Triggers WSS route resync on any transition."
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/wg/bootstrap": {
      "get": {
        "tags": ["Customer Tunnels"],
        "summary": "Download WireGuard .conf",
        "operationId": "wgBootstrapConf",
        "description": "Returns a ready-to-use WireGuard config file. The token is single-shot (24 h TTL) — consuming it marks it used.",
        "parameters": [
          {
            "in": "query",
            "name": "token",
            "required": true,
            "schema": { "type": "string", "minLength": 192, "maxLength": 192 },
            "description": "Single-shot bootstrap token (192 hex chars)."
          }
        ],
        "responses": {
          "200": {
            "description": "WireGuard configuration file",
            "content": {
              "text/plain": {
                "schema": { "type": "string" },
                "example": "[Interface]\nPrivateKey = ...\nAddress = 100.96.5.42/32\n\n[Peer]\nPublicKey = ...\nEndpoint = ams01.proxy.host.yt:51820\n"
              }
            }
          },
          "403": { "description": "Token invalid, expired, or already consumed", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "429": { "description": "Rate limit exceeded", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "500": { "description": "Internal error", "content": { "text/plain": { "schema": { "type": "string" } } } }
        }
      }
    },
    "/api/wg/install.sh": {
      "get": {
        "tags": ["Customer Tunnels"],
        "summary": "Download bash installer",
        "operationId": "wgInstallScript",
        "description": "Returns a bash script that installs `wireguard-tools`, fetches `/api/wg/bootstrap`, and creates a systemd unit. Designed for `curl … | sudo bash`. Pass `-- remove` to tear down an existing installation.\n\n```bash\n# Install\ncurl 'https://proxy.host.yt/api/wg/install.sh?token=<token>' | sudo bash\n\n# Remove\ncurl 'https://proxy.host.yt/api/wg/install.sh?token=<token>' | sudo bash -s -- remove\n```",
        "parameters": [
          {
            "in": "query",
            "name": "token",
            "required": true,
            "schema": { "type": "string", "minLength": 192, "maxLength": 192 },
            "description": "Bootstrap token — will be consumed by the script when it calls `/api/wg/bootstrap`."
          }
        ],
        "responses": {
          "200": {
            "description": "Bash installer script",
            "content": { "text/x-shellscript": { "schema": { "type": "string" } } }
          },
          "400": { "description": "Missing or malformed token", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "429": { "description": "Rate limit exceeded", "content": { "text/plain": { "schema": { "type": "string" } } } }
        }
      }
    },
    "/api/wg/status": {
      "get": {
        "tags": ["Customer Tunnels"],
        "summary": "WireGuard peer status",
        "operationId": "wgPeerStatus",
        "description": "Returns whether the peer has an active WireGuard handshake. Polled by the panel UI for a live connection indicator.",
        "parameters": [
          {
            "in": "query",
            "name": "token",
            "required": true,
            "schema": { "type": "string", "minLength": 192, "maxLength": 192 }
          }
        ],
        "responses": {
          "200": {
            "description": "Peer status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "enum": ["pending", "active"], "example": "active" },
                    "last_handshake": {
                      "oneOf": [
                        { "type": "string", "format": "date-time" },
                        { "type": "null" }
                      ],
                      "example": "2026-06-24T10:00:00Z"
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Missing or malformed token", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "404": { "description": "Unknown token", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "429": { "description": "Rate limit exceeded", "content": { "text/plain": { "schema": { "type": "string" } } } }
        }
      }
    },
    "/api/node/wg/peers": {
      "get": {
        "tags": ["Node Agent"],
        "summary": "Pull WireGuard peer list",
        "operationId": "nodePeersPull",
        "description": "Called by `hpg-node-agent` every ~30 s. Returns the full list of customer WireGuard peers the node should configure.",
        "security": [{ "nodeAuthToken": [] }],
        "parameters": [
          {
            "in": "query",
            "name": "node_token",
            "required": false,
            "schema": { "type": "string" },
            "description": "Alternative to `Authorization: Bearer`. One of the two must be present."
          }
        ],
        "responses": {
          "200": {
            "description": "Peer list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "peers": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "pubkey": { "type": "string", "example": "abc123...==" },
                          "allowed_ip": { "type": "string", "example": "100.96.5.42/32" },
                          "status": { "type": "string", "enum": ["active", "pending", "revoked"], "example": "active" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Missing token", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "403": { "description": "Invalid token", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "503": { "description": "Database not ready", "content": { "text/plain": { "schema": { "type": "string" } } } }
        }
      }
    },
    "/api/node/wg/handshakes": {
      "post": {
        "tags": ["Node Agent"],
        "summary": "Report WireGuard handshakes (legacy)",
        "operationId": "nodeHandshakeReport",
        "description": "Agent posts latest handshake timestamps. Superseded by `/api/node/wg/stats` but kept live for rolling upgrades.",
        "security": [{ "nodeAuthToken": [] }],
        "parameters": [
          { "in": "query", "name": "node_token", "required": false, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reports": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "pubkey": { "type": "string", "example": "abc123...==" },
                        "last_handshake": { "type": "string", "format": "date-time" }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": { "description": "Accepted" },
          "401": { "description": "Missing token", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "403": { "description": "Invalid token", "content": { "text/plain": { "schema": { "type": "string" } } } }
        }
      }
    },
    "/api/node/wg/stats": {
      "post": {
        "tags": ["Node Agent"],
        "summary": "Report WireGuard peer stats",
        "operationId": "nodeStatsReport",
        "description": "Full stats report from `wg show <iface> dump` — supersedes `/api/node/wg/handshakes`. Includes rx/tx byte counters, endpoint, and optional node-level forwarding diagnostics. Panel stores reset-safe byte deltas and updates `wstunnel_healthy`, triggering a Caddy route resync on any health transition.",
        "security": [{ "nodeAuthToken": [] }],
        "parameters": [
          { "in": "query", "name": "node_token", "required": false, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "stats": {
                    "type": "array",
                    "items": { "$ref": "#/components/schemas/WgPeerStats" }
                  },
                  "node": { "$ref": "#/components/schemas/NodeDiagnostics" }
                }
              },
              "example": {
                "stats": [
                  {
                    "pubkey": "abc123...==",
                    "last_handshake": 1719216000,
                    "rx_bytes": 1048576,
                    "tx_bytes": 524288,
                    "endpoint": "203.0.113.5:51820"
                  }
                ],
                "node": {
                  "ip_forward_enabled": true,
                  "forward_policy_drop_detected": false,
                  "docker_rules_installed": true,
                  "firewall_backend": "nftables",
                  "mtu": 1420,
                  "listen_port": "51821",
                  "last_setup_error": "",
                  "wstunnel_healthy": true
                }
              }
            }
          }
        },
        "responses": {
          "204": { "description": "Accepted" },
          "400": { "description": "Malformed body", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "401": { "description": "Missing token", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "403": { "description": "Invalid token", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "503": { "description": "Database not ready", "content": { "text/plain": { "schema": { "type": "string" } } } }
        }
      }
    },
...
    "/api/v1/services/{id}/routes": {
      "get": {
        "tags": ["Services"],
        "summary": "List routes for a service",
        "operationId": "serviceRoutes",
        "description": "Lists all routes attached to the service. Use `status` to track SSL provisioning progress. Clients can only query their own services.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": {
            "description": "Route list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "routes": { "type": "array", "items": { "$ref": "#/components/schemas/RouteObject" } }
                  }
                }
              }
            }
          },
          "401": { "description": "Auth required", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "403": { "description": "Not your service", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "404": { "description": "Service not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "429": { "description": "Rate limit exceeded", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  }
}